Separate betting in game module
