+++ oa_mod/code/game/g_combat.c
@@ -86,7 +86,7 @@ void AddScore(gentity_t* ent, vec3_t origin, int score) {
         return;
     }
     // no scoring during pre-match warmup
-    if ((level.warmupTime) || (g_gameStage.integer != PLAYING)) {
+    if ((level.warmupTime) || !isMatchTime()) {
         return;
     }
     //No scoring during intermission
